Methoden aus ProcedureController verschieben und alte deprecaten
